App.js-destructuring
