+++ src/App.js
@@ -1,52 +1,14 @@
-import React, { useState, useRef } from 'react';
+import React from 'react';
 import './App.css';
-import { Auth } from './components/auth/Auth';
-import Cookies from 'universal-cookie';
-import Chat from './components/chat/Chat';
-import { signOut } from 'firebase/auth';
-import { auth } from './config/firebase';
+import MainApp from './MainApp';
 
-const cookies = new Cookies();
-
-function App () {
-  const [isAuth, setIsAuth] = useState(cookies.get("auth-token"));
-  const [room, setRoom] = useState(null);
-
-  const roomInputRef = useRef(null)
-
-  const signUserOut = async () => {
-    await signOut(auth)
-    cookies.remove("auth-token")
-    setIsAuth(false)
-    setRoom(null)
-  };
-
-  if (!isAuth) {
-    return (
-      <div>
-        <Auth setIsAuth={setIsAuth} />
-      </div>
-    );
-  }           
-    
+const App = () => {
   return (
-    <>
-      {room ? (
-       <Chat room={room} />
-      ) : (
-        <div className='room'>
-          <label> room name</label>
-          <input ref={roomInputRef} />
-          <button onClick={ () => setRoom(roomInputRef.current.value)}>
-            Chat</button>
-        </div>
-      )}
-
-      <div className='sign-out'>
-        <button onClick={signUserOut}>Sign Out</button>
-      </div>
-    </>
+    <div>
+      <h1>Chat Application</h1>
+      <MainApp />
+    </div>
   );
-}
+};
 
 export default App;
